src/main/java/de/innovationhub/prox/modules/project/application/project/usecase/FindAllProjectsHandler.java
package de.innovationhub.prox.modules.project.application.project.usecase;

import de.innovationhub.prox.modules.commons.application.ApplicationComponent;
import de.innovationhub.prox.modules.project.domain.project.Project;
import de.innovationhub.prox.modules.project.domain.project.ProjectRepository;
import java.util.List;
import java.util.stream.StreamSupport;
import lombok.RequiredArgsConstructor;

@ApplicationComponent
@RequiredArgsConstructor
public class FindAllProjectsHandler {
  private final ProjectRepository projectRepository;

  public List<Project> handle() {
    return StreamSupport.stream(projectRepository.findAll().spliterator(), false).toList();
  }
}
